Capstone_Selenium_Pytest/requirements.txt
selenium
pytest
pytest-xdist
openpyxl
allure-pytest
webdriver-manager
configparser